+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/restapi/config/SwaggerConfig.kt
@@ -25,7 +25,6 @@ class SwaggerConfig {
     return GroupedOpenApi.builder()
       .group("API (outside of openApi)")
       .packagesToScan(
-        "uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.controller.statistics",
         "uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.controller.pni",
         "uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.controller.controllers",
       )
