+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/restapi/controller/controllers/AdminController.kt
@@ -1,11 +1,11 @@
-package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.controller
+package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.controller.controllers
 
+import io.swagger.v3.oas.annotations.Operation
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.service.ReferralService
 
-// Admin endpoints not visible via Swagger
 @RestController
 @RequestMapping("admin")
 class AdminController(
@@ -14,6 +14,11 @@ class AdminController(
 
   // Endpoint to update the cache in the person table
   // this should sparingly as it updates all the people in the database using the latest data from DPS.
+  @Operation(
+    tags = ["Admin"],
+    summary = "endpoint to update the cache in the person table. " +
+      "This should sparingly as it updates all the people in the database using the latest data from DPS.",
+  )
   @PostMapping("/person/updateAll")
   fun updatePersonCache() {
     referralService.updateAllPeople()
